Fixes for creating a porfile with billing information
Error out when sending a request fails for better error handling.

Add "company" as a field to billing information.

Make a customer profile paymentprofiles an actual list.

Swap payment profiles' billTO and Payment so that the json is generated
in that order.... apparently the Authorize.Net API cares about that.

Make logging through an interface so programs that use this library can
supply their own loggers.